client/scripts/requirements.txt
python-magic
pyOpenSSL
ndg-httpsclient
pyasn1
